ServerCLientImpl/fl_client/requirements.txt
# fl_client/requirements.txt
flwr==1.8.0
tensorflow-cpu==2.13.0
numpy==1.24.3
rich==13.7.1